Fork current driver/monitor framework
Existing driver/monitor bases and Avalon classes are copied to
new location.
This patch just duplicates the files without any changes so
all changes made to the files are clear:

* `cocotb/drivers/__init__.py` -> `cocotb/driver.py`
* `cocotb/monitors/__init__.py` -> `cocotb/monitor.py`
* `cocotb/drivers/avalon.py` -> `cocotbext/avalon/driver.py`
* `cocotb/monitors/avalon.py` -> `cocotbext/avalon/monitor.py`

Only Avalon code will be taken as example for porting.
Porting of other classes is left to maintainers of that code.